ItemCard 하드코딩된 데모 데이터를 동적 데이터로 교체 필요 : fix : flutter analyze 오류 수정 https://github.com/TEAM-ROMROM/RomRom-FE/issues/97

+++ lib/services/apis/item_api.dart
@@ -141,7 +141,7 @@ class ItemApi {
       onSuccess: (responseData) {
         // 응답이 int형이거나 Map형일 수 있으므로 안전하게 처리
         if (responseData is int) {
-          predictedPrice = responseData as int;
+          predictedPrice = responseData;
         } else if (responseData is Map<String, dynamic>) {
           predictedPrice = responseData['data'] ?? 0;
         } else {
